CAL-415 Addressd Sonarqube bug findings (#536)
* CAL-415 Addressd Sonarqube bug findings

* CAL-415 PR comments

* checkstyles

* addressed checkstyle errors

* resolved nsili itest failure

+++ catalog/imaging/imaging-transformer-chipping/src/main/java/org/codice/alliance/imaging/chip/transformer/AbstractJpegService.java
@@ -129,7 +129,7 @@ abstract class AbstractJpegService implements JpegService {
   }
 
   private double getTotalNumberOfPixels(BufferedImage bufferedImage) {
-    return bufferedImage.getWidth() * bufferedImage.getHeight();
+    return (double) bufferedImage.getWidth() * bufferedImage.getHeight();
   }
 
   private double getNumberOfBits(byte[] jpegData) {
