+++ catalog/video/video-mpegts-stream/src/main/java/org/codice/alliance/video/stream/mpegts/netty/BitReader.java
@@ -74,7 +74,7 @@ public class BitReader {
    * @return number of readable bits
    */
   public long readableBits() {
-    return byteBuf.readableBytes() * 8 + (currentByte != null ? currentIndex + 1 : 0);
+    return (long) byteBuf.readableBytes() * 8 + (currentByte != null ? currentIndex + 1 : 0);
   }
 
   /**
